Остановка(с синхронизацией) потоков
