+++ Tests/MailSender.ConsoleTest/Program.cs
@@ -1,114 +1,18 @@
 ﻿using System;
-using System.Globalization;
-using System.Threading;
 
 namespace MailSender.ConsoleTest
 {
     class Program
     {
-        public static void CheckThread()
-        {
-            var current_thread = Thread.CurrentThread;
-            Console.WriteLine("Поток \"{0}\"(id:{1}) запущен",
-                current_thread.Name, current_thread.ManagedThreadId);
-        }
-
         static void Main(string[] args)
         {
-            //System.Threading.Thread
-            Thread current_thread = Thread.CurrentThread;
-            current_thread.Name = "Главный поток";
+            ThreadTests.Start();
 
-            CheckThread();
-
-            //System.Diagnostics.Process
-            //Environment.ProcessorCount   // Кол-во процессоров
-
-            //var clock_thread = new Thread(new ThreadStart(ClockUpdater));
-            var clock_thread = new Thread(ClockUpdater);
-            clock_thread.Priority = ThreadPriority.Highest;
-            clock_thread.Name = "Поток часов";
-            clock_thread.IsBackground = true;
-            clock_thread.Start();
-
-            var message = "Hello World!";
-            //var printer_thread = new Thread(new ParameterizedThreadStart(Printer));
-            var printer_thread = new Thread(Printer);
-            printer_thread.Name = "Поток принтера";
-            printer_thread.IsBackground = true;
-            //printer_thread.Start(message);
-
-            var printer2_thread = new Thread(() => Printer(message));
-            printer2_thread.Name = "Поток принтера 2";
-            //printer2_thread.Start();
-
-            var printer3 = new Printer3(message);
-            var printer3_thread = new Thread(printer3.Print);
-            printer3_thread.Name = "Поток принтера 3";
-            printer3_thread.Start();
 
 
             Console.ReadLine();
         }
-
-        private static void ClockUpdater()
-        {
-            CheckThread();
-
-            while (true)
-            {
-                //Console.Title = DateTime.Now.ToString(CultureInfo.InvariantCulture);
-                Console.Title = DateTime.Now.ToString("dd:MM:yyyy HH:mm:ss.ffff");
-                Thread.Sleep(100);
-            }
-        }
-
-        private static void Printer(object obj)
-        {
-            CheckThread();
-
-            var message = (string) obj;
-
-            for (var i = 0; i < 20; i++)
-            {
-                Console.WriteLine("id:{0} - msg:\"{1}\"", Thread.CurrentThread.ManagedThreadId, message);
-                Thread.Sleep(100);
-            }
-
-            Console.WriteLine("Поток {0} завершён", Thread.CurrentThread.ManagedThreadId);
-        }
-
-        private static void Printer(string Message)
-        {
-            CheckThread();
-
-            for (var i = 0; i < 20; i++)
-            {
-                Console.WriteLine("id:{0} - msg:\"{1}\"", Thread.CurrentThread.ManagedThreadId, Message);
-                Thread.Sleep(100);
-            }
-
-            Console.WriteLine("Поток {0} завершён", Thread.CurrentThread.ManagedThreadId);
-        }
     }
 
-    internal class Printer3
-    {
-        private string _Message;
-
-        public Printer3(string Message) => _Message = Message;
-
-        public void Print()
-        {
-            Program.CheckThread();
-
-            for (var i = 0; i < 20; i++)
-            {
-                Console.WriteLine("id:{0} - msg:\"{1}\"", Thread.CurrentThread.ManagedThreadId, _Message);
-                Thread.Sleep(100);
-            }
-
-            Console.WriteLine("Поток {0} завершён", Thread.CurrentThread.ManagedThreadId);
-        }
-    }
+   
 }
